backend/dating/requirements/base.txt
django==1.10.*
psycopg2==2.6.2
Pillow==3.3.*
django-allauth==0.27.*
django-ipware==1.1.*
django-widget-tweaks==1.4.*
python-dateutil==2.5.*
redis==2.10.5
celery==3.1.24
aiopg==0.12.0
asyncio-redis==0.14.3
websockets==3.2
uvloop==0.5.4
rules==1.1.1
django-solo==1.1.2
django-autofixture==0.12.1
django-reversion==2.0.8
django-debug-toolbar==1.6
geoip2==2.4.2
